chore(#389): skills refactor
