+++ src/components/CharacterSheet/SkillsPage.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import { Character } from '../../types/Character';
 import { SkillsBonus } from '../../types/Character/Bonus';
 import Size from '../../types/Character/Size';
-import Skills, { Skill } from '../../types/Character/Skills';
+import { Skills, Skill } from '../../types/Character/Skill';
 import { StatSet } from '../../types/Character/Stats';
 import Header from '../helpers/Header';
 import { CheckCell, InputCell } from '../helpers/InputCell';
