+++ src/types/Character/Skill/Skill.ts
@@ -1,0 +1,55 @@
+enum Skill {
+    acrobatics = 'acrobatics',
+    appraise = 'appraise',
+    athletics = 'athletics',
+    bluff = 'bluff',
+    alchemy = 'alchemy',
+    blacksmithing = 'blacksmithing',
+    cooking = 'cooking',
+    inscription = 'inscription',
+    jewelcrafting = 'jewelcrafting',
+    leatherworking = 'leatherworking',
+    tailoring = 'tailoring',
+    technology = 'technology',
+    woodworking = 'woodworking',
+    diplomacy = 'diplomacy',
+    disarm = 'disarm',
+    disguise = 'disguise',
+    escape = 'escape',
+    handleAnimal = 'handleAnimal',
+    heal = 'heal',
+    intimidate = 'intimidate',
+    linguistics = 'linguistics',
+    arcane = 'arcane',
+    dungeons = 'dungeons',
+    engineering = 'engineering',
+    geography = 'geography',
+    history = 'history',
+    nature = 'nature',
+    nobility = 'nobility',
+    planes = 'planes',
+    religion = 'religion',
+    perception = 'perception',
+    act = 'act',
+    comedy = 'comedy',
+    dance = 'dance',
+    keyboard = 'keyboard',
+    percussion = 'percussion',
+    sing = 'sing',
+    speech = 'speech',
+    string = 'string',
+    wind = 'wind',
+    butchery = 'butchery',
+    enchanting = 'enchanting',
+    fishing = 'fishing',
+    herbalism = 'herbalism',
+    lumbering = 'lumbering',
+    mining = 'mining',
+    skinning = 'skinning',
+    ride = 'ride',
+    senseMotive = 'senseMotive',
+    sleight = 'sleight',
+    stealth = 'stealth',
+}
+
+export { Skill };
